Fixes from code review
Fixed "paddingBottom": "NaN%" for TimesImage
Moved styles to own subdirectory in src
Used styleguide for fonts
