+++ packages/video/src/styles/video.styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  overlay: {
+    position: "absolute",
+    top: 0,
+    left: 0,
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center"
+  }
+});
